+++ tests/test_utils/test_string_utils.py
@@ -1,6 +1,9 @@
+import pytest
+
 from quickbase_client.utils.string_utils import id_from_iso_string
 from quickbase_client.utils.string_utils import make_unique_var_name
 from quickbase_client.utils.string_utils import make_var_name
+from quickbase_client.utils.string_utils import normalize_hostname
 from quickbase_client.utils.string_utils import normalize_unicode
 from quickbase_client.utils.string_utils import parse_realm_and_app_id_from_url
 
@@ -78,6 +81,21 @@ class TestParseRealmAndAppIdFromUrl:
         realm, app_id = parse_realm_and_app_id_from_url(url)
         assert realm == "dicorp.quickbase.com"
         assert app_id == "bqx7xre7a"
+
+
+class TestNormalizeHostname:
+    @pytest.mark.parametrize(
+        "example",
+        [
+            "https://foo.quickbase.com",
+            "foo.quickbase.com",
+            "https://foo.quickbase.com/",
+            "http://foo.quickbase.com/",
+            "foo.quickbase.com/",
+        ],
+    )
+    def test_normalizes(self, example):
+        assert normalize_hostname(example) == "foo.quickbase.com"
 
 
 class TestNormalizeUnicode: